fix: satisfy pkg-file-count gate for buffered pause tests
Move engine buffered-pause coverage into pkg/factory/engine/bufferedpausetests
and fold session-scoped service tests into existing runtime-mode/helper test
files so pkg/factory/engine and pkg/service stay within the 15-file limit.
