--- pkg/factory/engine/engine_buffered_pause_submit_test.go
@@ -1,123 +1,0 @@
-package engine
-
-import (
-	"context"
-	"testing"
-
-	"github.com/portpowered/infinite-you/pkg/factory/subsystems"
-	"github.com/portpowered/infinite-you/pkg/interfaces"
-	"github.com/portpowered/infinite-you/pkg/petri"
-)
-
-func TestSubmitWhileAutomaticTicksPaused_AcceptsAndBuffersUntilResume(t *testing.T) {
-	n := buildTestNet()
-	marking := petri.NewMarking("test-wf")
-	sub := &mockSubsystem{group: subsystems.Scheduler}
-
-	paused := true
-	engine := NewFactoryEngine(n, marking, []subsystems.Subsystem{sub}, WithAutomaticTicksPaused(func() bool {
-		return paused
-	}))
-
-	request := interfaces.WorkRequest{
-		RequestID: "request-paused-submit-001",
-		Type:      interfaces.WorkRequestTypeFactoryRequestBatch,
-		Works: []interfaces.Work{{
-			Name:       "paused-submit",
-			WorkTypeID: "task",
-			TraceID:    "trace-paused-submit",
-		}},
-	}
-	result, err := engine.SubmitWorkRequest(context.Background(), request)
-	if err != nil {
-		t.Fatalf("SubmitWorkRequest while paused: %v", err)
-	}
-	if !result.Accepted {
-		t.Fatalf("submit result accepted = false, want true")
-	}
-	if result.RequestID != request.RequestID {
-		t.Fatalf("submit result requestID = %q, want %q", result.RequestID, request.RequestID)
-	}
-
-	assertNoTokensInPlace(t, engine, "task:init")
-	if err := engine.Tick(context.Background()); err != nil {
-		t.Fatalf("Tick while paused: %v", err)
-	}
-	assertNoTokensInPlace(t, engine, "task:init")
-	if sub.callCount != 0 {
-		t.Fatalf("subsystem callCount = %d, want 0 while paused", sub.callCount)
-	}
-
-	paused = false
-	if err := engine.Tick(context.Background()); err != nil {
-		t.Fatalf("Tick after resume: %v", err)
-	}
-	snap := engine.GetMarking()
-	tokens := (&snap).TokensInPlace("task:init")
-	if len(tokens) != 1 {
-		t.Fatalf("tokens in task:init = %d, want 1 after resume", len(tokens))
-	}
-	if tokens[0].Color.TraceID != "trace-paused-submit" {
-		t.Fatalf("token traceID = %q, want trace-paused-submit", tokens[0].Color.TraceID)
-	}
-	if sub.callCount != 1 {
-		t.Fatalf("subsystem callCount = %d, want 1 after resume", sub.callCount)
-	}
-
-	repeated, err := engine.SubmitWorkRequest(context.Background(), request)
-	if err != nil {
-		t.Fatalf("duplicate SubmitWorkRequest: %v", err)
-	}
-	if repeated.Accepted {
-		t.Fatal("duplicate submit should be idempotent no-op")
-	}
-	if err := engine.Tick(context.Background()); err != nil {
-		t.Fatalf("Tick after duplicate submit: %v", err)
-	}
-	snap = engine.GetMarking()
-	if len((&snap).TokensInPlace("task:init")) != 1 {
-		t.Fatalf("token count after duplicate submit = %d, want 1", len((&snap).TokensInPlace("task:init")))
-	}
-}
-
-func TestWakeForPendingProcessing_SignalsBufferedSubmissionAfterPausedWake(t *testing.T) {
-	n := buildTestNet()
-	marking := petri.NewMarking("test-wf")
-	sub := &mockSubsystem{group: subsystems.Scheduler}
-
-	paused := true
-	engine := NewFactoryEngine(n, marking, []subsystems.Subsystem{sub}, WithAutomaticTicksPaused(func() bool {
-		return paused
-	}))
-
-	if _, err := submitWorkRequests(context.Background(), engine, []interfaces.SubmitRequest{{
-		WorkTypeID: "task",
-		TraceID:    "trace-paused-wake",
-	}}); err != nil {
-		t.Fatalf("SubmitWorkRequest: %v", err)
-	}
-
-	// Simulate a paused wake attempt consuming the submit signal without processing.
-	if err := engine.Tick(context.Background()); err != nil {
-		t.Fatalf("Tick while paused: %v", err)
-	}
-	assertNoTokensInPlace(t, engine, "task:init")
-
-	paused = false
-	engine.WakeForPendingProcessing()
-	if err := engine.Tick(context.Background()); err != nil {
-		t.Fatalf("Tick after resume wake: %v", err)
-	}
-	snap := engine.GetMarking()
-	if len((&snap).TokensInPlace("task:init")) != 1 {
-		t.Fatalf("buffered submission was not reachable after paused wake and resume")
-	}
-}
-
-func assertNoTokensInPlace(t *testing.T, engine *FactoryEngine, placeID string) {
-	t.Helper()
-	snap := engine.GetMarking()
-	if got := len((&snap).TokensInPlace(placeID)); got != 0 {
-		t.Fatalf("tokens in %s = %d, want 0 while paused", placeID, got)
-	}
-}
